Generate item-level norms for age, sex, language profile and item dominance.

Generate item-level norms for age, sex, language profile and item dominance.

ml_norms(
  participants = NULL,
  responses = NULL,
  item = NULL,
  language = c("Catalan", "Spanish"),
  type = c("understands", "produces"),
  age = c(0, 100),
  lp = c("Bilingual", "Monolingual", "Other"),
  sex = c("Female", "Male"),
  category = NULL,
  .width = 0.95
)

Arguments

participants
Participants data frame, as generated by ml_participants. If NULL (default), ml_participants is run.

responses
Responses data frame, as generated by ml_responses. If NULL (default), ml_responses is run.

item
Character string indicating the item to compute norms for. If left NULL (by default) norms will be computed for all items. You can check the available items in the pool data set running data("pool").

language
Character string indicating the language to compute vocabulary norms for: "catalan" and/or "spanish"

type
Character string indicating the vocabulary type to compute norms for. Takes "understands" and/or"produces" (defaults to both).

age
Numeric vector of length two (min-max) indicating the age range of participants to compute norms for.

lp
character string indicating the language profile of participants to compute norms for: "Bilingual", "Monolingual", "Other" (defaults to all).

sex
character string indicating the sex of participants to compute norms for. Takes "Female" and/or "Male" (defaults to both).

category
character string indicating the semantic/functional category(ies) to include items from. See available categories in the pool data set by running data("pool").

.width
Numeric values ranging from 0 to 1 (not included) indicating the confidence level of confidence intervals (defaults to 0.95).

Value

A data frame with the proportion of participants in the sample that understand or produce the items indicated in item, along with the standard error and confidence interval of the estimation.

Details

Estimated proportions and corresponding standard errors and confidence intervals are computed following Gelman et al. (2020) 's adjustments to account for zero- and one-inflation (see functions prop_adj, prop_adj_se, and prop_adj_ci).

References

Gelman A, Hill J, Vehtari A (2020). Regression and other stories. Cambridge University Press.

Examples

ml_norms(item = "cat_casa", type = "understands", age = c(20, 24))

#> Updating data...

#> BL-Long-2 updated

#> BL-Lockdown updated

#> # A tibble: 12 x 15
#>       te item   language age_bin type  lp    category item_dominance label   yes
#>    <int> <chr>  <chr>      <dbl> <chr> <chr> <chr>    <chr>          <chr> <int>
#>  1   591 cat_c~ Catalan       22 unde~ Bili~ Outside  L1             casa     24
#>  2   591 cat_c~ Catalan       24 unde~ Bili~ Outside  L1             casa     21
#>  3   591 cat_c~ Catalan       22 unde~ Bili~ Outside  L2             casa     18
#>  4   591 cat_c~ Catalan       24 unde~ Bili~ Outside  L2             casa      5
#>  5   591 cat_c~ Catalan       22 unde~ Mono~ Outside  L1             casa     19
#>  6   591 cat_c~ Catalan       24 unde~ Mono~ Outside  L1             casa     12
#>  7   591 cat_c~ Catalan       22 unde~ Mono~ Outside  L2             casa      5
#>  8   591 cat_c~ Catalan       24 unde~ Mono~ Outside  L2             casa      2
#>  9   591 cat_c~ Catalan       22 unde~ Other Outside  L1             casa      2
#> 10   591 cat_c~ Catalan       24 unde~ Other Outside  L1             casa      2
#> 11   591 cat_c~ Catalan       22 unde~ Other Outside  L2             casa      1
#> 12   591 cat_c~ Catalan       24 unde~ Other Outside  L2             casa      1
#> # ... with 5 more variables: n <int>, proportion <dbl>, se <dbl>,
#> #   ci_lower <dbl>, ci_upper <dbl>