Retrieve and update local and/or remote data from formr

Retrieve and update local and/or remote data from formr

ml_responses(
  participants = NULL,
  runs = c("BL-Long2", "BL-Lockdown"),
  longitudinal = "all",
  update = TRUE
)

Arguments

participants
Participants data frame, as generated by ml_participants. If NULL (default), ml_participants is run.

runs
formr runs to update

longitudinal
Should longitudinal data be included? If "all" (default), all responses (including repeated measures) are included. If "no", participants with more than one responses to the questionnaire (regardless of the version) are excluded. If "first", only the first response of each participant is included. If "last", only the last response of each participant is included. If "only", only responses with repeated measures are included.

update
Should data be recovered from last update, or should it be updated now?

Value

A data frame (actually, a tibble) containing participant's responses to each item, along with some session-specific metadata. The output includes the following variables:

id

a character string indicating a participant's identifier. This value is always the same for each participant, so that different responses from the same participant share the same id.

id_exp

a character string indicating a participant's identifier in the context of the particular study in which the participant was tested and invited to fill in the questionnaire. This value is always the same for each participant within the same study, so that different responses from the same participant in the same study share id_exp The same participant may have different id_exp across different studies.

id_db

a character string with five digits indicating a participant's identifier in the database from the Laboratori de Recerca en Infància at Universitat Pompeu Fabra. This value is always the same for each participant, so that different responses from the same participant share the same id_db.

time

a numeric value indicating how many times a given participant has been sent the questionnaire, regardless of whether they completed it or not.

code

a character string identifying a single response to the questionnaire. This value is always unique for each response to the questionnaire, even for responses from the same participant.

study

a character string indicating the study in which the participant was invited to fill in the questionnaire. Frequently, participants that filled in the questionnaire came to the lab to participant in a study, and were then invited to fill in the questionnaire later. This value indicates what study each participant was tested in before being sent the questionnaire.

version

a character string indicating what version of the questionnaire a given participant filled in. Different versions may contain a different subset of items, and the administration instructions might vary slightly (see formr questionnaire templates in the GitHub repository). Also, different versions were designed, implemented, and administrated at different time points (e.g., before/during/after the COVID-related lockdown).

time_stamp

a date value (see lubridate package) in yyyy/mm/dd format indicating the date in which participants responded to the last item of their questionnaire response. Note: some participants took longer to complete the questionnaire since they started filling items in.

language

a character string indicating the language the item response belongs to: Catalan if item in Catalan), Spanish if item in Spanish.

item

character string indicating the item identifier (e.g., spa_mesa). This value is unique for each item. Responses to the same item from different participants are linked by the same item value.

response

integer indicating the participant's response to a give item: 1 if No (the participant does not understand or prooduce the word), 2 if Understands (the participants understands the word), or 3 if Understands and Says (the participant understands and produces the item).

date_birth

a date value (see lubridate package) in yyyy/mm/dd format indicating participants birth date.

age

a numeric value indicating the number of months elapsed since participants' birth date until they filled in the last item of their questionnaire response.

sex

a character string indicating participants' biological sex, as reported by the parents.

postcode

a character string indicating participants' household postcode.

edu_parent1

a character string indicating the educational attainment of one of the parents/caretakers.

edu_parent2

a character string indicating the educational attainment of the other parent/caretaker, if any.

doe_spanish

a numeric value ranging from 0 to 1 indicating participants' daily exposure to Spanish, as estimated by parents/caretakers. This value aggregates participants' exposure to any variant of Spanish (e.g., European and American Spanish).

doe_catalan

a numeric value ranging from 0 to 1 indicating participants' daily exposure to Catalan, as estimated by parents/caretakers. This value aggregates participants' exposure to any variant of Catalan (e.g., Catalan from Majorca or Barcelona).

doe_others

a numeric value ranging from 0 to 1 indicating participants' daily exposure to languages other than Spanish or Catalan, as estimated by parents/caretakers, aggregating participants' exposure to all those other languages (e.g., Norwegian, Arab, Swahili).

randomisation

a character string indicating the specific list of items a participant was assigned to. Only applies in the case of short versions of multilex, such as BL-Short, BL-Short-2 or BL-Lockdown, where the list of items was partinioned into several versions.

doe_spanish_lockdown

a numeric value ranging from 0 to 1 indicating participants' daily exposure to Spanish during the COVID-19 lockdown, as estimated by parents/caretakers. This value aggregates participants' exposure to any variant of Spanish (e.g., European and American Spanish).

doe_catalan_lockdown

a numeric value ranging from 0 to 1 indicating participants' daily exposure to Catalan during the COVID-19 lockdown, as estimated by parents/caretakers. This value aggregates participants' exposure to any variant of Catalan (e.g., Catalan from Majorca or Barcelona).

doe_others_lockdown

a numeric value ranging from 0 to 1 indicating participants' daily exposure to languages other than Spanish or Catalan during the COVID-19 lockdown, as estimated by parents/caretakers, aggregating participants' exposure to all those other languages (e.g., Norwegian, Arab, Swahili).

dominance

a character string indicating the language of highest exposure ("Catalan" or "Spanish"), as reported by parents. If exposure is identical for both language, "Catalan" is assigned.

Details

This function generates a data frame with participant's responses to each item, along with some session-specific metadata. It takes participants (the output of ml_participants) and runs (a character vector that can take zero, one, or multiple of the following values: "formr2", "formr-short", "formr-lockdown") as arguments. Only responses from the versions indicated in runs input will be updated. For the rest, data will be retrieved from the their last update.

Author

Gonzalo Garcia-Castro